requirements.txt
aiohttp==3.5.4
aiopg==0.16.0
async-timeout==3.0.1
attrs==19.1.0
chardet==3.0.4
idna==2.8
multidict==4.5.2
psycopg2-binary==2.8.3
yarl==1.3.0
astroid==2.2.5
autopep8==1.4.4
colorama==0.4.1
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
pycodestyle==2.5.0
pylint==2.3.1
PyYAML==5.1.2
six==1.12.0
typed-ast==1.4.0
wrapt==1.11.2
